Update badge performance (#344)
* Make unauthed badge loads batched

* Don't build tooltip until badges are set

* Run prettier
