+++ AntGameUI/src/antGame/User/BadgeService.js
@@ -1,6 +1,5 @@
 import ResultCache from "../Helpers/ResultCache";
-import { GetBatchBadges, GetUserBadges } from "./UserService";
-import AuthHandler from "../Auth/AuthHandler";
+import { GetBatchBadges } from "./UserService";
 
 class BadgeService {
   constructor() {
@@ -28,30 +27,13 @@ class BadgeService {
 
   async fetchBadges() {
     if (!this.fetchingBadge && this.badgesToFetch.length > 0) {
-      if (AuthHandler.loggedIn) {
-        if (this.fetchTimeout) clearTimeout(this.fetchTimeout);
-        this.fetchTimeout = setTimeout(this.getBatchBadges.bind(this), 10);
-      } else {
-        this.getSingleBadge();
-      }
+      if (this.fetchTimeout) clearTimeout(this.fetchTimeout);
+      this.fetchTimeout = setTimeout(this.getBatchBadges.bind(this), 10);
     }
   }
 
-  async getSingleBadge() {
-    this.fetchingBadge = true;
-
-    const userID = this.badgesToFetch.shift();
-    const badges = await GetUserBadges(userID);
-    const resolveFunction = this.badgesCache.getValue(userID).resolveFunc;
-    resolveFunction(badges);
-
-    this.fetchingBadge = false;
-    this.fetchBadges();
-  }
-
   async getBatchBadges() {
-    if (!AuthHandler.loggedIn || this.fetchingBadge)
-      throw new Error("Illegal call to getBatchBadges");
+    if (this.fetchingBadge) throw new Error("Illegal call to getBatchBadges");
     this.fetchingBadge = true;
 
     const userIDList = [...this.badgesToFetch];
